Add MultiScreenCaptureAllowedForUrls for Lacros
In a previous CL, the same policy was introduced on Ash. This CL
implements the `IsGetAllScreensMediaAllowed` oracle for Lacros.
In order to do so, the signature of `IsGetAllScreensMediaAllowed`
needs to be changed to accept a result-callback instead of returning
a boolean.

tests as we cannot make the Ash binary start up with a certain policy
value. This would be required for a complete end2end test as the policy
value is only cached at Ash startup. Adding a test interface to set
the policy on the Ash side would defeat the purpose of an end2end test.
We have covered that the Ash side works properly as we have end2end
test for Ash, which also go through the crosapi interface (not IPC).

Bug: b/40216442
Change-Id: I4b17e63c73df9cc1c0c9d10a93b01f2dfc044d20
Design: go/multiscreencapture-iwa-policy
Low-Coverage-Reason: HARD_TO_TEST We are mocking the crosapi for system
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5398120
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Commit-Queue: Simon Hangl <simonha@google.com>
Cr-Commit-Position: refs/heads/main@{#1286456}