Update test user constants and skip unncessary combinations in test
This CL updates constants in login_manager_mixin.h to align
with other mixins that set up users for testing.

Update FamilyUserDeviceMetrics test to correctly handle
presence of the test user before login in some tests.

This is done in preparation to change the default value of
recovery opt-in for consumer users.

Bug: b:325279785
Change-Id: I035f6e70434a0c23f6a765293f564be8b11f0f97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444837
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Denis Kuznetsov <antrim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286475}