[Read aloud] Add voice selection into toolbar keyboard navigation
In Reading mode toolbar with read aloud, arrow key navigation was
skipping over the voice selection menu. This CL adds it into the
navigation sequence to make it keyboard accessible.

The approach here is to move the toolbar button out of the
voiceSelectionMenu element, to align better with the pattern used by
the other toolbar menus.

We also considered keeping the structure as-is, to avoid the anti-pattern of having a parent call child methods. However, this would
have added complexity to identifying the set of focusable toolbar
buttons. This one would be inside shadowRoot and not returned by
querySelectorAll. Plus, keyboard event targets would be the
voiceSelectionMenu and not the toolbar button. We decided it was cleaner to keep a consistent pattern across the menus for now and look at reducing anti patterns in a more global way in future.

Bug: 40927698
Change-Id: Ice41b6cf33b79861cb1871dfa5ae6d533db6025b
Fixed: 332300395
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5446320
Reviewed-by: Lauren Winston <lwinston@google.com>
Reviewed-by: Eitan Goldberger <eitang@google.com>
Commit-Queue: Shari Trewin <trewin@google.com>
Cr-Commit-Position: refs/heads/main@{#1286517}