[ios] SetUpList long press menu item to toggle Tips Notifications
This change modifies the SetUpList long-press context menu so that when
the user has already opted-in, it presents a menu item to allow turning
off the notifications. If the user has not opted-in yet, a menu item
will be presented to allow opting-in (this is existing behavior).

Fixed: 330745105
Change-Id: I3976e753189accc01b0a6c3fffc6b45c4a27841b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444668
Commit-Queue: Scott Yoder <scottyoder@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286535}