[Customize] Stop button overflow for color scheme mode
When text was too long to fit all of the sections into the button, it
would just make the segmented-button-option wider than the overall
segmented-button (see bug for example video).

To handle this, min-width is set to 0 so the button-option will stop
using content as the min-width, causing the button-option to be larger
than the button. Also, as discussed in the bug, overflow text will be
hidden with ellipsis. Then, full text will be visible via tooltip on
hover.

Screenshots:
1 long word overflow: http://screenshot/7M6JFwjv6n6xhov
multi-word overflow: http://screenshot/BTLSYeY5yyr2XXQ

Bug: 41481914
Change-Id: Ied17c40e0c0b0e6a31408aa5eaa76a3fa8b65914
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445927
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Riley Tatum <rtatum@google.com>
Cr-Commit-Position: refs/heads/main@{#1286555}