[Tab Group Parity] Do not notify new group creation on undone closure
On longpress tab group creation, tabs added to the group will have a
parent tab associated with it. This becomes problematic on undoing a
closure since it will notify didCreateNewGroup when that should not be
the case. This added boolean flag prevents that from triggering.

Bug: 332366113
Change-Id: I8c45a8b4702efe3efd6bb1c92b203fc4d906ba54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444673
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Commit-Queue: Brandon Fong <bjfong@google.com>
Cr-Commit-Position: refs/heads/main@{#1286563}