Add the option to specify video plane updates in graphics coordinates.
This is part of the work to get cast on starboard building out of
chromium. See go/moving-cwr-to-chromium for more information on the
high-level goal.

When setting video plane bounds in Starboard (via SbPlayerSetBounds),
the coordinates must be in terms of the graphics resolution, NOT the
physical screen resolution. For example, the graphics resolution on a
commercial TV may be 1080p for a 4k TV. In this case, the video player's
bounds must be set relative to the 1080p resolution.

Without this CL, VideoPlaneController converts coordinates to the
physical screen resolution, causing problems in Starboard.

Bug: b/333571227
Change-Id: I3d4ef66e3c592770fda01bc99baec79b3dc43158
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5440456
Reviewed-by: Shawn Quereshi <shawnq@google.com>
Commit-Queue: Antonio Rivera <antoniori@google.com>
Reviewed-by: Yuchen Liu <yucliu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286576}