Use the first child included in layout for sizing
When a view is removed, the parent view will ask to relayout before
removing the child view from their children lists. When the

When SimpleGridLayout attempts to cache the children view size, the
removed child view will technically still be in the list of children
views of the host view, but it will have a size 0 since it will be
removed.  To avoid the edge case where the first child view is removed
and a bad layout is calculated, take into consideration the first child
that will actually be included in the layout.

Bug: b:333338805
Change-Id: Iaeba771dbf08859584914cb5a6636cd558f050a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5446054
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286625}