oak birch: decouple birch bar and overview layout
This cl decouples the birch bar bounds change and overview layout
such that the appearance of birch bar will not affect the positions
of overview windows unless the user intentionally show or hide the
bar from context menu.

When the birch bar is enabled, we leave a space for birch bar at the
bottom of the Overview and the windows are positioned in the rest
available space. When birch bar shows or hides, we only update the
wallpaper cropping area and feedback button position to make them
always above the top of birch bar.

The demo is shown in the bug comment.

Bug: b:332558522
Change-Id: Ib598fee6cce7ab4934c364f51c1494dd596ac8a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5448858
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiaodan Zhu <zxdan@chromium.org>
Reviewed-by: Michele Fan <michelefan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286631}