Pass IsolationInfo to ContentBrowserClient::WillCreateURLLoaderFactory()
Adds a new `const net::IsolationInfo& isolation_info` parameter to
`ContentBrowserClient::WillCreateURLLoaderFactory()`.

This is done in order to pass the information about the top level
frame origin to embedders, such as Chrome. The immediate use-case
example for this is https://crrev.com/c/5360379.

The isolation info is only provided for untrusted factories which are
not allowed to set `ResourceRequest::TrustedParams::isolation_info` in
their network requests. Untrusted factories are the ones that are shared
with untrusted processes. This is controlled by the
`URLLoaderFactoryParams::is_trusted` parameter when a factory is
created.

To update existing callers, this CL used the following algorithm:

- determine whether a factory is going to be trusted by looking at
`URLLoaderFactoryParams::is_trusted`
- if trusted, pass `net::IsolationInfo()` signifying that the
isolation info is not fixed and that each request will likely provide
an isolation info in `ResourceRequest::TrustedParams::isolation_info`
- otherwise, pass into `WillCreateURLLoaderFactory()` the same
isolation info as being set in `URLLoaderFactoryParams::isolation_info`

Bug: 331803339
Change-Id: I7eeff21bd5649b4fe4c4d0cdab59b26ed4388f45
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5405301
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Auto-Submit: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Reviewed-by: mmenke <mmenke@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286642}