immersive fullscreen: fix oom on window restore
Restoring a fullscreen window that has an NSToolbar and an
NSTitlebarAccessoryViewController causes the key view loop to become
corrupted. The NSToolbar's view points to the
NSTitlebarAccessoryViewController's view in the key view loop. Under
certain circumstances this can cause an infinite loop leading to an OOM.

This CL calls `-[NSWindow recalculateKeyViewLoop]` on the browser window
after a restore. This fixes the key view loop setting the NSToolbar's
view `-nextKeyView` to nil.

Note: This bug stems from a workaround added because of FB12010731
(NSTitlebarAccessoryViewController fullScreenMinHeight ignored). Once
that bug is fixed this workaround can be removed.

For good completeness I have also reported this bug to Apple under
FB13724144 (restoring a fullscreen window that has an NSToolbar and an
NSTitlebarAccessoryViewController causes the key view loop to become
corrupted). Though we would prefer a fix for FB12010731.

Bug: 324812653
Change-Id: I813b049ec844e26c343395b7e99ff05808b3ffe8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445152
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Commit-Queue: Tom Burgin <bur@google.com>
Cr-Commit-Position: refs/heads/main@{#1286666}