[Telemetry] Support interactive routine in telemetry extension
Add a field `interaction` to the waiting state so that a routine can
ask callers to react to an interaction request. For interactions that
need to be replied, callers should invoke the `replyToRoutineInquiry`
method.

The implementation of `ReplyInquiry` in cros_healthd was done in
CL:5171546.

Bug: b/309781398
Test: unit_tests -ozone-platform=headless --gtest_filter="*Diagnostics*"
Test: browser_tests -ozone-platform=headless --gtest_filter="*Telemetry*"
Change-Id: I226f46f8f070595d7ef59c72d456a3e457681f23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5411712
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Justin Lulejian <jlulejian@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286675}