Render bounding box after the user makes a selection
Use the CSS Houdini Paint API to render a bounding box that
shows the user their selection. The corners of the bounding box are
rendered using the Paint API, which will allow us to more easily and
performantly animate the corners. In addition, the PostSelectionRenderer
is a separate element as it will be in charge of letting the user grab
a corner to adjust the region, and since there will be multiple ways to
make a selection, I decided to have one element handle rendering the
post selection.

Demo: https://screencast.googleplex.com/cast/NDc0NTcxMTQ4MzYxNzI4MHw4NzUzOWMyMC04Mw

Bug: b/332894403
Change-Id: I777ca459b9602643aa8faa818ad88adeac0dfd44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5429683
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1286694}