Add new UI for enterprise profile creation screen
This new UI has the basic functionality of the legacy one, but does
not yet support bringing browser data back. This will first be used
for the OIDC flow which does not use that.
This CL is required for initial testing of the feature, more tests
automated tested will come, and once the legacy dialog is removed,
that dialog's tests should work the same on the new dialog.

This CL also ensures that the management badge is shown on both the
legacy and updated screens when the intercepted account is a managed
account.

Screenshot: http://screen/8ainJRbAfaLWqHt
Screenshot dark mode: http://screen/335YqZ3cnr2JqDf

Bug: 332735128
Change-Id: I77540f4c4cb7e1ad75ae9a99ddb12c6aa44edd35
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5419171
Reviewed-by: Jack Yammine <jyammine@google.com>
Reviewed-by: Monica Basta <msalama@chromium.org>
Commit-Queue: Yann Dago <ydago@chromium.org>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286705}