Spanify and convert from BigEndianReader to SpanReader in //media
BigEndianReader is a subset of the functionality in SpanReader, so
it is being sunsetted and replaced with SpanReader.

While converting, we also spanify methods that were constructing
BigEndianReader from a pointer.

Add SpanReader::ReadCopy() as a replacement for
BigEndianReader::ReadBytes() when the code needs a copy of a byte
span instead of just a reference (span) as its output.

Have SpanReader::Skip() return an optional span of the bytes that
were skipped over. This allows boolean chaining to skip and pass
the resulting span to other boolean-returning functions without
having to pull out the span first and resize it, and do size checks
manually. This matches the Skip API of SpanWriter.

Bug: 40284755
Change-Id: I8ad311ddf81c5df1d5c92f687dbca32886f6a979
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5439778
Reviewed-by: Frank Liberato <liberato@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286733}