[FedCM] Keep button flow modal open during use other account flow
Currently, the button flow modal is hidden when the use other account
flow is triggered. This patch changes it such that the modal remains
open. Selecting an account or clicking on "Cancel" resolves the promise
and leaves the pop-up open. But clicking on the "Use a different
account" button again while a pop-up is open closes the pop-up to reopen
it. This is so that the new pop-up appears in front of the window and is
re-centred.

Bug: 333414306
Change-Id: I0929ef30e1187d4425a447716d1b32f6f7690829
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5445764
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Zachary Tan <tanzachary@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286746}