Fix crash in sharing menu
-[NSMenu setAutoenablesItems:NO] is dangerous. -[NSMenuDelegate
menuNeedsUpdate:] is called before a menu is shown, but not
necessarily before a menu is searched for menu shortcut keys.
Therefore, if you disable the autoenabling of items, you run the
risk of the enable state of getting out of sync with what should be
allowed.

Therefore, remove that call and allow the autoenabling of items.
Implement the NSMenuItemValidation protocol, so that however the menu
is accessed, by showing or by shortcut key, the enable state is always
correct.

Fixed: 333424895
Change-Id: I5aeacf3bc2514f9c5ba20008b686b5380927cea4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5449679
Reviewed-by: Leonard Grey <lgrey@chromium.org>
Auto-Submit: Avi Drissman <avi@chromium.org>
Commit-Queue: Leonard Grey <lgrey@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286769}