[tab group v2] Saved tab group migration from v1 to v2
Since existing saved tab groups in v1 are considered unpinned in v2.
A one time migration is needed to set the first 4 SavedTabGroup
to pinned as user turns on the feature for the first time.
A profile pref is used to record of state of whether the user has
been migrated to v2.

Bug: 333759143
Change-Id: I6974529cb434c81959050d2690d1d787f784acba
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5444262
Reviewed-by: Darryl James <dljames@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Commit-Queue: Yuheng Huang <yuhengh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286806}