Open searchbox suggestion URLs in the LensOverlay side panel.
Additionally, updates <enter> key and click behavior to close the dropdown. This affects the NTP searchbox (realbox) as well, which historically just depended on the web contents it was hosted in navigating to the destination URL and didn't bother updating the UI. But this new behavior more closely matches the behavior of other searchboxes (for example, the google.com searchbox, which closes the dropdown before navigating).

Bug: b:332787629
Change-Id: I26fe9826a03c4d6262cf7a37f91930aba262f662
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5448814
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1286813}