mahi: Scroll to top/bottom on applicable state changes
Depending on the state of the mahi panel, will scroll to bottom or top
to show the appropriate contents.

1. Navigate to Q&A view -> scroll to bottom
2. New Q&A added -> scroll to bottom
3. Navigate to Summary view -> scroll to top
4. New summary view contents -> scroll to top

Pixel tests were updated to take a screenshot of the scroll view instead
of the question answer view, since the latter might be bigger than the
visible scroll view's viewport and would capture unwanted areas, such
as the panel's header. Restricting the screenshot to the scroll view
will ensure we only account for the visible portion of the view.

Bug: b:330413997
Change-Id: Ic03e64da5b3346ea70eaa850e82c8b9be2657b88
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5422348
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Reviewed-by: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286833}