[Compose] Add redo to Compose dialog and saved state model
The current model uses the `undo_states_` stack to track the past states of the dialog associated with responses. In this model, there is a notion of a `most_recent_ok_state_`, which is tied to the last response received and is not saved into the stack until a newer response is received. Thus, `undo_states_` only owns ComposeStates preceding the current one, and is empty if there is only one saved response in the session.

In the proposed model, `undo_states_` is replaced by the `history_` vector. In this model, the state history owns all saved states, including the most recent one. New responses are pushed onto the history vector as they are received. Thus, the `most_recent_ok_state_` concept is removed. Position in the history is tracked by `history_current_index_`, which is used by CurrentState() to get the state at the current position in the history and by LastResponseState() for the last saved state in history tied to a response (relative to that current position).
The purpose of these two indices is to distinguish between saved states that represent a model response, and saved states that represent a user edit on a response. In the context of this CL (ignoring already landed editable results changes) the later type of saved state is not yet introduced, and so these functions are nearly identical.

The redo function, represented by a new button and new icons, is gated behind the refinements flag.

Bug: b/325341277
Change-Id: Iee2bc47c65cbb50b2068e8d4d0bf3422c72cc206
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5441513
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286850}