[Compose] Merge modifiers behind UI refinement flag
This change combines the Tone and Length enums currently defined in compose.mojom into a single StyleModifiers enum, and handles the corresponding refactor. It adds a new select containing all modifiers to the dialog, which replaces the existing Tone and Length menus when the ComposeUiRefinements flag is enabled.

Bug: b/325340923
Change-Id: Ic4afb821a48fceb4204760e019bf5931eee1d21d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5446065
Reviewed-by: Justin DeWitt <dewittj@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Trevor Perrier <perrier@chromium.org>
Commit-Queue: Anthony Cui <cuianthony@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286848}