Add lifecycle constraints to new PersistedTabData |from| API
Custom Tabs, uninitialized Tabs and destroyed Tabs should all
result in null when the |from| API is used. This is because
PersistedTabData is a UserData object, it's not possible to be used on
an uninitialized or destroyed Tab. So we need to account for these use
cases, including as data is passed between threads. PersistedTabData is
also not supported for Custom Tabs at all.

Bug: b:331413918
Change-Id: Iccb98e2a8b61d4c4b19fd9d414b3546795f05ca9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5450516
Reviewed-by: Brandon Wylie <wylieb@google.com>
Commit-Queue: David Maunder <davidjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286866}