bluetooth: Avoid dangling raw_ptr on channel close
BluetoothRfcommChannelMac and BluetoothL2capChannelMac maintain a
delegate that wraps the IOBluetooth channel. The delegate maintains a
pointer back to the Bluetooth*ChannelMac so it can forward events.

Bluetooth*ChannelMac will be destroyed once OnChannelClosed is received.
Ensure the delegate resets its pointer to the Bluetooth*ChannelMac so
it doesn't keep a dangling reference.

Change-Id: Ifaddfbdc5b38533d9137ca14097484b444891023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5450849
Reviewed-by: Alvin Ji <alvinji@chromium.org>
Commit-Queue: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286884}