Add test demonstrating PrinterQueryOop::UpdatePrintSettings() crash
Introduce a test that demonstrates how the browser can be left in an
inconsistent state if the renderer process crashes while handling the
rendering for printing.  This can cause a subsequent attempt to print
from the Print Preview to fail on a CHECK.

Add an extra observer to PrintViewManagerBase so that tests can track
when a print render frame has been deleted, such as when the renderer
process terminates.

Extend KillPrintRenderFrame class so that the renderer is killed when
it tries to dereference GetForwardingInterface(), which occurs when it
is processing the command to print.

This new test is initially disabled since it crashes on a CHECK.  Use
of EXPECT_CHECK_DEATH() didn't work here, as that interfered with the
test browser enough to cause an "Browser unresponsive" dialog to be
displayed.

Bug: 332512063
Change-Id: Ib161d0b5b9f9403aad39aa85a60af2d1d2f20c3f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5450234
Reviewed-by: Andy Phan <andyphan@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1286900}