Use Lacros omnibox search controller factory from client impl
This is slightly suboptimal. If the factory is not bound when the
provider class is constructed, the dedicated search controller is only
created when `GetController()` is called - which would be on the first
search. This might increase the latency of the first search.

Additionally, if the factory is _rebound_, a new controller is not
created. This is unlikely to happen, unless lacros-chrome crashes.

This will be addressed in a follow-up CL which observes when the factory
is bound.

Bug: b:326147929
Change-Id: I7a1a545e4df4c6f390bd134ee188aa57387ea8c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5450725
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Michael Cui <mlcui@google.com>
Auto-Submit: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1286958}